Create a detailed search function in the search bar
Create a selectValue state and change its value whenever an onChange event occurs on the select element.
Put the selectValue state in the searchValue atom so that the corresponding condition is requested to the api as well.
Add a variable to the value of the option element to resolve the warnings in React related to injection.
In addition, set the initial value of the searchValue atom to 'don't send' and process it in HomeContainer.js to prevent unnecessary api requests during the first rendering.